requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.23
numpy>=1.24
soundfile>=0.12
PyYAML>=6.0
JACK-Client>=0.5.4
